pyproject.toml
[tool.poetry]
name = "bitsplit"
version = "0.1.0"
description = ""
authors = ["Mohit Solanki <mohitsolanki619@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^2.1"
djangorestframework = "^3.9"
python-decouple = "^3.1"
dj-database-url = "^0.5.0"
django-extensions = "^2.1"
Pillow = "^5.3"
django-mptt = "^0.9.1"

[tool.poetry.dev-dependencies]
ptpython = "=0.41"
model_mommy = "^1.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
